requirements.txt
rapidfuzz==3.6.1
msal==1.25.0
msgraph-sdk==1.0.0
python-dotenv==1.0.0
requests==2.31.0
azure-identity==1.15.0
motor==3.3.2
pymongo[srv]==4.6.1
fastapi==0.105.0
boto3==1.34.5
python-multipart==0.0.6
uvicorn==0.25.0
authlib==1.2.1
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0
